requirements.txt
django>=5.0
django-storages[azure]>=1.14
django-bootstrap5>=23.3
MySQLClient>=2.1
Pillow>=10.1
